!sched/clock: remove CONFIG_SYSTEM_TIME64 and always use 64-bit time
The 32-bit system clock has a limited range (~497 days) and the
configuration knob is no longer worth the complexity given that
practically every modern target already enables it.  Make 64-bit
time_t/clock_t/sclock_t/nuttx_time_t the only supported flavor.

Specifically:
  - Drop the SYSTEM_TIME64 Kconfig option and its dependent
    PERF_OVERFLOW_CORRECTION/HRTIMER guards in sched/Kconfig.
  - Remove every #ifdef CONFIG_SYSTEM_TIME64 branch in headers
    (include/{sys/types.h,limits.h,inttypes.h,nuttx/clock.h,
    nuttx/fs/hostfs.h}) and core code paths
    (sched/clock/clock.h, drivers/power/pm/pm_procfs.c,
    drivers/rpmsg/rpmsg_ping.c, fs/procfs/fs_procfsuptime.c,
    libs/libc/wqueue/work_usrthread.c,
    arch/avr/src/avrdx/avrdx_timerisr_tickless_alarm.c).
  - Strip CONFIG_SYSTEM_TIME64=y from every board defconfig.
  - Update Documentation/guides/rust.rst accordingly.

+++ arch/arm/src/efm32/efm32_rtc_burtc.c
@@ -466,7 +466,7 @@ int up_rtc_settime(const struct timespec *tp)
 
   /* Compute Burtc offset because we cannot reset counter */
 
-  val = (((uint64_t)tp->tv_sec) * CONFIG_RTC_FREQUENCY) + \
+  val = (tp->tv_sec * CONFIG_RTC_FREQUENCY) + \
         (tp->tv_nsec / (NSEC_PER_SEC / CONFIG_RTC_FREQUENCY));
 
   if (val < cnt_reg)
