+++ arch/risc-v/src/esp32c3-legacy/esp32c3_tim_lowerhalf.c
@@ -386,7 +386,7 @@ static int esp32c3_timer_settimeout(struct timer_lowerhalf_s *lower,
 
   /* Set the timeout */
 
-  ESP32C3_TIM_SETALRVL(priv->tim, (uint64_t)timeout);
+  ESP32C3_TIM_SETALRVL(priv->tim, timeout);
 
   return ret;
 }
