+++ arch/xtensa/src/esp32s3/esp32s3_oneshot.c
@@ -288,8 +288,8 @@ int esp32s3_oneshot_start(struct esp32s3_oneshot_s *oneshot,
 
   /* Retrieve the duration from timespec in microsecond */
 
-  timeout_us = (uint64_t)ts->tv_sec * USEC_PER_SEC +
-               (uint64_t)(ts->tv_nsec / NSEC_PER_USEC);
+  timeout_us = ts->tv_sec * USEC_PER_SEC +
+               (ts->tv_nsec / NSEC_PER_USEC);
 
   /* Verify if it is a multiple of the configured resolution.
    * In case it isn't, warn the user.
@@ -459,7 +459,7 @@ int esp32s3_oneshot_current(struct esp32s3_oneshot_s *oneshot,
 
   ESP32S3_TIM_GETCTR(oneshot->tim, usec);
 
-  *usec = *usec * (uint64_t)oneshot->resolution;
+  *usec = *usec * oneshot->resolution;
 
   return OK;
 }
