+++ boards/arm/stm32/b-g474e-dpow1/src/stm32_smps.c
@@ -466,7 +466,7 @@ static int smps_start(struct smps_dev_s *dev)
     {
       pwrerr("ERROR:  Can not achieve timc pwm "
              "freq=%" PRIu32 " if fclk=%" PRIu64 "\n",
-             (uint32_t)TIMC_PWM_FREQ, (uint64_t)fclk);
+             (uint32_t)TIMC_PWM_FREQ, fclk);
       ret = -EINVAL;
       goto errout;
     }
@@ -483,7 +483,7 @@ static int smps_start(struct smps_dev_s *dev)
     {
       pwrerr("ERROR:  Can not achieve timd pwm "
              "freq=%" PRIu32 " if fclk=%" PRIu64 "\n",
-             (uint32_t)TIMD_PWM_FREQ, (uint64_t)fclk);
+             (uint32_t)TIMD_PWM_FREQ, fclk);
       ret = -EINVAL;
       goto errout;
     }
